requirements.txt
colorama==0.4.6
slack-sdk==3.27.1
python-dotenv==1.0.1
certifi==2024.2.2 